test: merge test indexer helper
